Added logic behind adding completed doors and windows, now the added window/door is added to the table. Changed the long values of doors and windows objects to integer.

+++ src/main/java/objects/BQDoor.java
@@ -4,14 +4,14 @@ import java.util.Map;
 
 public class BQDoor extends Door{
 
-    private Map<String, Long> materialNameAndQuantity;
+    private Map<String, Integer> materialNameAndQuantity;
 
-    public BQDoor( String doorName, Map<String, Long> materialNameAndQuantity) {
+    public BQDoor( String doorName, Map<String, Integer> materialNameAndQuantity) {
         super(doorName);
         this.materialNameAndQuantity = materialNameAndQuantity;
     }
 
-    public Map<String, Long> getMaterialNameAndQuantity() {
+    public Map<String, Integer> getMaterialNameAndQuantity() {
         return materialNameAndQuantity;
     }
